749c752 <     if ($i == 0 or $song->bpm($i) != $song->bpm($i-1)) { >     if ($i == 0 or $i == 1 or $song->bpm($i) != $song->bpm($i-1)) {
$song->bpm( 0 ) is undefined. When $i == 1 this spat an error ($i - 1 = 0), so I added a check for it.

===
870c873
<     my ($self,$color,$family,$size,$text,$x,$y) = @_;
>     #my ($self,$color,$family,$size,$text,$x,$y) = @_;

I couldn't work out what this was trying to achieve. It's at the end of a function, and none of the values listed are returned. Hashing it out didn't seem to do any harm, and got rid of the warning about redefining $self.

===
Cleaned up a few extraneous "my" statements that were causing warnings.